dns type/class uint16

+++ zonefileutil/const.go
@@ -22,7 +22,7 @@ const (
 	DNS_CLASS_ANY = 255
 )
 
-var DnsIntTypes map[int]string = map[int]string{
+var DnsIntTypes map[uint16]string = map[uint16]string{
 	//	DNS_TYPE_ALL:   "ALL",
 	DNS_TYPE_A:     "A",
 	DNS_TYPE_NS:    "NS",
@@ -35,7 +35,7 @@ var DnsIntTypes map[int]string = map[int]string{
 	DNS_TYPE_SRV:   "SRV",
 	DNS_TYPE_ANY:   "ANY",
 }
-var DnsStrTypes map[string]int = map[string]int{
+var DnsStrTypes map[string]uint16 = map[string]uint16{
 	//	DNS_TYPE_ALL:   "ALL",
 	"A":     DNS_TYPE_A,
 	"NS":    DNS_TYPE_NS,
@@ -49,12 +49,12 @@ var DnsStrTypes map[string]int = map[string]int{
 	"ANY":   DNS_TYPE_ANY,
 }
 
-var DnsIntClasses map[int]string = map[int]string{
+var DnsIntClasses map[uint16]string = map[uint16]string{
 	//DNS_CLASS_ALL: "ALL",
 	DNS_CLASS_IN:  "IN",
 	DNS_CLASS_ANY: "ANY",
 }
-var DnsStrClasses map[string]int = map[string]int{
+var DnsStrClasses map[string]uint16 = map[string]uint16{
 	//DNS_CLASS_ALL: "ALL",
 	"IN":  DNS_CLASS_IN,
 	"ANY": DNS_CLASS_ANY,
